fix: :bug: Improved separation of concerns for search page
